Add PixelExt trait
